requirements.txt
amqp==5.0.9
asgiref==3.4.1
asn1crypto==0.24.0
attrs==21.4.0
Automat==0.6.0
autopep8==1.6.0
billiard==3.6.4.0
Brotli==1.0.9
cached-property==1.5.2
celery==5.1.2
certifi==2018.1.18
chardet==3.0.4
charset-normalizer==2.0.10
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.3.7
ConfigArgParse==1.5.3
configobj==5.0.6
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.1.4
dataclasses==0.8
defusedxml==0.7.1
Deprecated==1.2.13
distro-info===0.18ubuntu0.18.04.1
dj-database-url==0.5.0
Django==3.2.11
django-cors-headers==3.10.1
django-debug-toolbar==3.2.4
django-filter==21.1
django-redis==5.2.0
django-silk==4.2.0
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.4.0
djoser==2.1.0
docopt==0.6.2
drf-nested-routers==0.93.4
Flask==2.0.2
Flask-BasicAuth==0.2.0
Flask-Cors==3.0.10
flower==1.0.0
gevent==21.12.0
geventhttpclient==1.5.3
gprof2dot==2021.2.21
greenlet==1.1.2
gunicorn==20.1.0
httplib2==0.9.2
humanize==3.13.1
hyperlink==17.3.1
idna==2.6
importlib-metadata==1.7.0
incremental==16.10.1
iniconfig==1.1.1
itsdangerous==2.0.1
itypes==1.2.0
Jinja2==3.0.3
keyring==10.6.0
keyrings.alt==3.0
kombu==5.1.0
language-selector==0.1
locust==2.6.0
MarkupSafe==2.0.1
model-bakery==1.4.0
msgpack==1.0.3
mysqlclient==2.0.3
netifaces==0.10.4
oauthlib==3.1.1
packaging==21.3
PAM==0.4.2
Pillow==8.4.0
pluggy==1.0.0
prometheus-client==0.12.0
prompt-toolkit==3.0.24
psutil==5.9.0
py==1.11.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycodestyle==2.8.0
pycrypto==2.6.1
Pygments==2.11.2
PyGObject==3.26.1
PyJWT==1.7.1
pyOpenSSL==17.5.0
pyparsing==3.0.6
pyserial==3.4
pytest==6.2.5
pytest-django==4.5.2
pytest-watch==4.2.0
python-apt==1.6.4
python-dateutil==2.8.2
python-debian==0.1.32
python3-openid==3.2.0
pytz==2021.3
pyxdg==0.25
PyYAML==3.12
pyzmq==22.3.0
redis==4.1.1
requests==2.27.1
requests-oauthlib==1.3.0
requests-unixsocket==0.1.5
roundrobin==0.0.2
SecretStorage==2.3.1
service-identity==16.0.0
six==1.11.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.2
ssh-import-id==5.7
systemd-python==234
toml==0.10.2
tornado==6.1
Twisted==17.9.0
typing_extensions==4.0.1
ufw==0.36
uritemplate==4.1.1
urllib3==1.22
vine==5.0.0
watchdog==2.1.6
wcwidth==0.2.5
Werkzeug==2.0.2
whitenoise==5.3.0
wrapt==1.13.3
zipp==3.6.0
zope.event==4.5.0
zope.interface==4.3.2
